requirements.txt
monai==1.3.0
numpy==1.26.4
pandas==2.2.1
pillow==10.3.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.13.0
torch==2.2.0
torch-tb-profiler==0.4.3
torchaudio==2.2.0
torchinfo==1.8.0
torchmetrics==1.2.0
torchvision==0.17.0